nand_base.c: Handle bitflip count of last page correctly
In drivers/mtd/nand/nand_base.c, in nand_do_read_ops(), if the last
page that is read in the loop has a bitflip count > 0, do not return
this directly after the loop. We still need to check stats.failed and
we should return the highest bitflip count in any case.