Use dynamic NAND page layout for fsvybrid
The array of NAND layouts assumed NAND flashes with 64 bytes OOB. This
will make problems if we use NAND flashes with larger OOB areas in the
future.

The new code now uses a dynamic NAND layout that is computed at runtime
when the OOB size is already known. Then it will work for all OOB sizes.

Conflicts:

	drivers/mtd/nand/fsl_nfc_fus.c