[fsvybrid] Fix ETH clock on armStoneA5
New armStoneA5 revisions use an external oscillator for the ETH clock
to work around a hardware limitation of the Vybrid CPU. NBoot detects
when this oscillator is mounted and sets feature bit FEAT2_RMIICLK_CKO1
then. Add code to detect this bit in case of armStoneA5 and configure
ETH clock as input instead of output then.

Remove CONFIG_FS_VYBRID_PLL_ETH because we have automatic detection for
internal or external clock in all cases now and don't need this manual
switch anymore.