Improve authenticate functionality
we improved the whole authenticate structure so that is nearly the same
like in NBoot. Some files doesnt exist anymore because we dont need them
like ivt.c and ivt.h. HAB.c/h are moved to small letter hab.c/h. The
check methods in ivt.c are merged to checkboot.c.