Fix MAC address handling on fsvybrid
The single MAC address for WLAN on CubeA5 has to be set in ethaddr, not
eth1addr. This also has to be handled differently when exporting the MAC
addresses to the bdinfo device tree node in ft_board_setup().

This all takes place in board/F+S/fsvybrid/fsvybrid.c.