Add support for fat_write for fsimx6ul_mmc
We need the fat write command to update kernel and dts via
update script.
So we have to add the original fat driver, which supports fat write.

Also fix the fat Makefile so that fat.o does not get build when
fat_write.o is build, because it includes fat.c.