Add display test with j070wvt lvds
Add display  test for j070wvtc0211 LVDS display.

Put the video related code in a extra file

We need to use the panel infrastrcture for the tc358764 to work.

In order to do so we have tu manipulate the sec_mipi_dsim driver ore the
timing will get computet wrong.