Update display support for adp-mipi2lvds1 rev120
ATTENTION: There is still a bug for the J070WVTC1601W display. When the
display initializes, it can happen (~10%) that the first pixel lane
switches between the beginning and the end of the display. The color
also switches when the pixel lane is at the end.
It is unclear, where the fault for the initialization is located. It
could be, that the issue is resolved when updating to a newer U-Boot
version. This issue is not present in Kernel, so for further analysis a
comparison between U-Boot and Kernel might help.