Update video initialization
ATTENTION: The initialization of the MIPI2LVDS adapter is mostly copied
from the kernel. However, there still is a bug in the timing of the
initialization that sometimes the first line of the display will appear
at the end of the display. This causes the display to flicker with an
inverted color.
Reassess this issue in a later UBoot version. (in 2020.04 or 2021.04)