Add Board-Configs for the PicoCoreMX8MMr2
Since there are diffrences in the Hardware, the PicoCoreMX8MMr2 is
looked at as a diffrent board variant. Because of this it needs its own
Board-CFGs.

The r2 Variants are consistent with the regular Variants of the same
Number. Thus the Board-CFGs can be copied with minor changes. But since
not every Variant that exists as a r2 Board in the fsimx8mm-Y2921.04
release we add those that are relevant to the r2 Variant for
completeness.

Also add the Board Type so the SPL knows what  kind of Board it is
running on, when one of the new configs is loaded.