Improve ethernet test
The ethenet test only worked for a singe ethernet device so far, because
the "not-ethact" ethernet did not get initialized properly while booting
and the phy device dit not get set.
Now we set the ethact device for each tested device and initialize it
before we read out the connected phy.