Change CONFIG_EXTRA_ENV_SETTINGS
These changes allow images to boot directly from memory
(boot partition).
In addition, images are validated internally for Secureboot.

This changes the boot concept of the LSDK by loading, validating and
executing U-Boot scripts from memory. The boot scripts in the LSDK
contain the commands to be executed for validating and booting a Linux
system for the reference design boards.

Here the concept is now adapted for GAL boards as follows.
When =>boot is executed:

1. Search for a valid bootpartition in the memory (MMC/USB).

2. Search for a "fsgal_boot.scr" in the boot partition, validated
when secureboot is activ and execute it.

3. If no script is found, images (kernel & DTB) are searched for.

4. If images are found, they are validated when secureboot is activ and
executed.

5. If no image is found, the boot process continues like in the LSDK, or
as defined in config_distro_bootcmd.h which the LSDK uses.