Add the board-cfg to bdinfo
We want to be able to read the board-cfg and board-id from linux.
For this we improve the bdinfo functionality, so it will automatically
add the board-cfg and board-id to the linux bdinfo device-tree node.
Linux will have to implement it's own routine to read out the DTS node
and add it to the bdifo sysfs entries.

We have to move two functions form fsimace.c to fs_image_common.c so we
can use it in fs_fdt_common.c

Also we have to backport the fdt_overlay_apply_node from 2021.03.