Add possibility to run GPIO tests by name
Sometimes we need to run GPIO tests independently of uclass devices, but
not with all the other gpio tests.

For example if we need to run an ethernet GPIO test on a board without
ethernet, there will be no ethernet uclass we can use.

Introduce a new function test_gpio_name, which can be used to run
different GPIO tests by name.

The name has to be the same as in the gpio device tree node e.g for
"AUDIO"

gpiotest-AUDIO
out-pins-AUDIO= ...
out-gpios-AUDIO= ...
in-pins-AUDIO= ...
in-gpios-AUDIO= ...

Also,  append "-IO" to the IO tests in the output, to make it more
cleare which test is functional and which test is IO.