Symbolic changes in GAL devicetree
Add the alias ethX for each Ethernet interface.
This alias is used to assign the ethXaddr environment variable to the
correct interface.

In addition:
The Ethernet nodes got the same label as in the Linux DTS. Since labels
are stored in the __symbols__ node within the / node, this can be used
to reference the exact same node between Linux and UBoot device tree.
By using the __symbols__ node we can change the properties of each node
in the same way as in DTS or DTSI files.

This theoretically allows us to use fs-gal1-dts for all GAL boards.
For this reason it was renamed from fs-gal1 to fs-gal.