Add board features definition and board setup
During U-Boot startup the board type, the board revision and the board
configuration can be recognized in UBoot. For this getter methods were
defined, which read OUIDR3 & OUIDR4 and determine, which board it
concerns here.

Based on this information, a list of available features are defined for
each specific board.
Each feature represence the usage of a device within LS1028 SoC.

These features are used to edit the UBoot devicetree within
board_fix_fdt(); and the Linux-devicetree within ft_board_setup();

Furthermore:
UBoot is reading the Mac-Address and a total number of used
Mac-Addresses and configures the Network interfaces and ethXaddr
environment Variables