Add dram test and processor_info to bdinfo
We want to output all testresults in linux but the dramtest and
processor_info are only available in uboot.
So we write the testresults to bdinfo in uboot and access them in linux
when printing the output.

This has also the advantage, that we can run the dram test befor loading
the initramfs into the ram or else the dramtest would damage the image.