Set Reset-Signals for Ethernet-Phys in board_init()
The GPIO signals for the PHY resets are cleared in board_init().

This is especially necessary for the Realtek Phy, as its reset is
connected to 1v8 with a pull-up to avoid disturbing the clock generator
during the power-on reset.

With the Realtek Phy, it has been found that a simple reset clear is not
sufficient to successfully start the PHY.

When a power-on reset of the PHY is performed, the power sequence first
sets the supply voltage of the PHY (3V3) and then 12.5ms later the reset
of the PHY by the active 1V8 voltage.

However, the power sequence of the Phys takes 100ms and a reset within
this time will result in the configured PHYAD not being read. Therefore,
the PHY must complete its power-up sequence once and then be reset to
read its address.