Set GPIOs for PCIe Connectors in board_init()
PWR-Switches for PCIe right-hand side and left-hand side are set to
enabled.
The VERSACLOCK-Output 6 and 7 are set to enabled.

Currently the PERST signal is not set. That could cause a PCIe device to
be not detected in U-Boot. The Linux-Kernel will set the proper
reset-signals when the system is booting.

Currently, a PCIe power enable affects the behaviour of the RGMII-Phy.
It loses its configured PHYAD[2:0]. The cause is assumed to be the
3V3_ETH power supply and is currently still being investigated.

For this reason, PCIe1_PWR_EN and PCIe2_PWR_EN are switched with a
small delay. At the end of board_init, the reset signals of the RGMII-PHY are
implemented.