set correct Reserve-Ram address in tfa_dram_init_banksize()
This commit fixes a bug in the current tfa_dram_init_banksize()
function.

The global variable resv_ram stores the start address for the reserved
memory area in U-Boot. This reserved area is used e.g. by NXP for
Ethernet purposes in some Layerscape SoCs.

The tfa_dram_init_banksize() function should set resv_ram to the highest
available ram address not reserved by U-Boot.

Currently, only addresses from DRAM region 1 are set.
This comes from the following if condition

if (gd->bd->bi_dram[1].size >
    board_reserve_ram_top(gd->bd->bi_dram[1].size)) {
	gd->arch.resv_ram = gd->bd->bi_dram[1].start +
		gd->bd->bi_dram[1].size -
		board_reserve_ram_top(gd->bd->bi_dram[1].size);
} else if (gd->bd->bi_dram[0].size >
	   board_reserve_ram_top(gd->bd->bi_dram[0].size)) {
	gd->arch.resv_ram = gd->bd->bi_dram[0].start +
		gd->bd->bi_dram[0].size -
		board_reserve_ram_top(gd->bd->bi_dram[0].size);
}

On the LS1028, DRAM_Region 1 is not used if the main memory is less than
2GiB. In this case, resv_mem must get an address from DRAM_Region0,
since the first 2GiB are addressed there.

On the LS1028, the return value of board_reserve_ram_top() is always
0x0. So no reserved area is needed in the U-Boot.

If bi_dram[1].size is also 0, the first if-condition is 0>=0, which is
true. Since we are now using less than 2GiB of RAM, the start address of
DRAM_Region1 is not necessarily defined. resv_ram is therefore set to
0x0.

This is a serious problem because when booting the Linux system,
ft_cpu_setup() is called in the board files for Layerscape SoCs.

When ft_cpu_setup() is executed, ls_gic_rd_tables_init() is called,
which needs resv_ram to determine its base address for the gic.

gic_lpi_base = ALIGN(gd->arch.resv_ram - GIC_LPI_SIZE, SZ_64K);

If resv_ram is 0x0, an invalid address is calculated for gic_lpi_base,
causing the CPU to crash.

This bug can be fixed with a corrected IF condition in
tfa_dram_init_banksize().