fs-gal.dtb: Add mmc_pwrseq for MMC-reset during reboot
A soft reset only turns off the voltages generated by the PMIC. Devices
powered by 3V3 are not reset. This includes the eMMC on the GAL boards.

The mmc-pwrseq-emmc driver can be used to perform a specified MMC reset
just before reset the U-Boot system.

This ensures that the BL1 can still read from the MMC after a soft
reset.