fs-gal.dtb: Reset-QPHY to prevent deadlock during softreset
QPHY uses a 2V5 as Supply-Voltage. The 2V5 are comming from PMIC.
During Soft-Reset, the PMIC disables the 2V5 output and enables it after
11.5ms.

It seems that the QPHY goes into an undefined state during the
soft-reset. This causes a deadlock in U-Boot during Ethernet
initialization. Therefore, this phy is reset in a controlled manner.