Improve fdt preparation during board setup
There are some changes to the feature definition and FDT preparation
during board setup.

Now the U-Boot can determine if a UART interface is used for RS485 or
RS232. If RS485 is used, the specific UART interface has a defined
property linux,rs485-enabled-at-boot-time to enable RS485 mode during
Linux boot.

Now, by default, U-Boot will enable all interfaces (except network
interfaces) that are used on specific boards such as GAL1. This means
that it is not possible to disable a device in the device tree if the
interface is available on the board. But it is always possible to enable
a device node of an interface that is not defined in the board features.

Fixed a bug in FEAT_GAL_ETH_INTERN_BASEX. Now it is possible to enable
enetc-port1 AND I2C when using internal BASEX connector on GAL boards.