Reset and Hold reset for PCIe Ports
Some PCIe devices use a pull-up or pull-down for the reset signal.
So it is not always a default, that the reset is set.

Some other devices need a clear reset before PCI initialization.

So we reset all PCIe devices in U-Boot in a controlled manner.
Linux will then clear the reset before initializing pci.