fix pwrseq for PCI devices
In order to use PCI devices with MSI or MSIX interrupts in Linux, the
devices must be successfully enumerated in U-Boot. The PCI enumeration
is done via pci_init();.

As soon as a PCI device is detected on a PCI-RC port, U-Boot configures
iommu and msi mappings in pci_layerscape_fixup.c for the device tree of
the Linux system.

Currently, the correct PWRSEQ could not be performed in U-Boot. This was
due to the fact that some PCI devices provide the PERST (reset) with a
pull-up resistor. So it must be ensured that the PERST signal is pulled
to ground before the power is switched on.

This should ensure that all PCI devices in the U-Boot are recognized and
the necessary memory mappings for booting into the linux systems are
defined.

For better code readability, board_init() would be revised and function
calls would be added to prepare external interfaces like pci,
ethernet-phy, usb, etc. for startup.