Adjustment for enviroment variables
Reason for the changes:
The env-variables console, rootfs and nfsroot cause Errors during boot
process.
The console was defined as "console=console=ttyLP1,115200 earlycon\0",
which does not works with netargs.

Aim is to make our FSNET_ENV compatiable with initial enviroment
variables from NXP.

Added definiton for nfsroot.

Fixed the confilicts in rootfs, console and bootargs.