enable MMC High Speed modes
The MMC card is compatible with HS200 and HS400. fsgal_tfa_defconfig now
defines HS200 and HS400 support and the devicetree for fsgal has been
adapted. The configuration now allows faster read and write speeds for
the MMC device.

For SD card interfaces UHS is not supported, so a no-1-8-v property has
been added to the esdhc node.