board:F+S:common:fs_bootrom: improve error handling
If an F&S Header Magic is found in an unexpected location, this is not
necessarily an error.

For example, there are scripts in the NBOOT that use the F&S HEADER
Magic in programming. Such a magic may also be recognized in
unfavourable positions in Data-Stream and is therefore not an error
while loading.