arm:dts:imx93: Add a per clock for lpuart3 to lpuart8
Add a per clock for lpuart3-8. Do it on the same way like NXP lpuart1.

Original NXP message is:
MLK-25904-1 arm: dts: imx93: add a per clock for LPUART1
When CLK is enabled, get_lpuart_clk_rate() needs to get a per clock of
lpuart, so that add a per clock for lpuart1