F+S:fsimx93:nboot: set properties in nboot-info
The BOARD-CFG typically provides properties such as spl-size,
nboot-size,
uboot-start, and uboot-size.

These values need to be set dynamically for two reasons:

Reason 1:
U-BOOT-INFO is at the end of nboot, so uboot-start depends on
nboot-size.

Reason 2:
This information cannot be included in the board CFG on the MMC via cmd
fsimage, as this would invalidate the hashes and signatures of the
container header.

This information is automatically determined by the SPL during the MMC
boot process. This data is now also stored in the board CFG in OCRAM.

Additionally:
The U-BOOT ENV is placed at the end of the MMC BOOTPART and uses the
secondary BOOTPART as a redundant copy. The new sizes and offsets are
also included in nboot-info.dtsi.
Each saveenv command sets both copies