board:F+S:fsimx93: edit DDR-TIMINGs for PICOCOREMX93
The DDR-CLOCK in Normal-Drive mode needs to be the exact half DDR-RATE
as of the Over-Drive Mode.

The DDR-CLOCK in normal drive mode must have exactly half the DDR rate
of the overdrive mode.

The ATF uses a bit in the DDR controller to set this frequency directly.

The timing parameters for normal drive must therefore be set for
1600MT/s.