board:F+S:fs_cntr_common: provide nboot-info and signature validation
During Boot via MMC/NAND, fs_handle_uboot() use the U-Boot offset in
flash to provide nboot-infos, like nboot-size, uboot-start, uboot-size.

The properties are required in OCRAM for the fsimage cmd.

To ensure a valid signature within an IMX_CONTAINER, the properties in
the BOARD-CFG cannot be determined by fsimage or compile time. Since the
SPL knows the offsets to the U-Boot when booting via seekable devices,
the known information can simply be stored here in the OCRAM.

---

fs_cntr_is_valid_signature() and fs_cntr_is_signed() can be used to
validate a IMX_CONTAINER image.

The signature is checked via AHAB. The SHA is checked directly at
"loadaddr" in RAM.