board:F+S:fs_cntr_common: Expect DRAM-FW dynamically during loading
The iMX93 used to expect the DRAM FW at a specific position in
DRAM-INFO. However, since no FW is required on the iMX8ULP, the load
process must not fail if the DRAM-FW is not found.

The changes allow the DRAM-INFO to be loaded for both architectures. The
DRAM-FW, if present, must be loaded as the first image. DRAM timings
must then be available.