board:F&S:common:fs_cntr_common: Workaround loading board-cfg for 8ulp
TODO:
Architectures like the imx8ulp need to use alternative sram memories to
store the board-cfg. Depending on the boot peripherals, the bootrom can
store data at any address. However, it turns out that USDHC0 cannot load
data directly into SRAM1 via Boot-ROM. This can be done via USB. It is
assumed that the RDC needs further configuration that is currently
unknown. MDA and MRC seems to have the required Configuration for the
Access.

A simple workaround is to check between Loadaddr and
CFG_FUS_BOARDCFG_ADDR and copy the valid binary with the ID of Cortex-A.