cmd:fsimage: fix board-id overwrite for cntr images
fs_image_find_board_cfg() checks whether a new board ID is available in
the ram. If this is the case and the user allows the board ID to be
overwritten, the compare ID is set.

The compare-id is used to find the correct board-cfg for the board with
fs_match_board_cfg().

However, the global variable "board-id" is not overwritten, but remains
the same for the time being.

The compare-id must be used so that the new board ID can now be saved in
the FSH ‘BOARD-ID’ in FLASH.

Therefore, fs_image_get_compare_id() is defined to generate a
description of the BOARD-ID, which can be used in FSHs.

This means that a new board ID can also be saved in the EMMC for CNTR
images.