board:F+S:common:fs_bootrom: ensure alignment during the search for FSH
Previously, it was assumed that a non-aligned FSH in the data stream was
unusual. However, in rare cases, the alignment may not match. This
special case must be taken into account, especially with critical
headers such as Board ID or Board and DRAM CFG.

bootrom_find_fshdr_stream() will now ensure the correct alignment of a
FSH within the global buffer.