WinIoT: Add fs_deviceinfo driver
The fs_deviceinfo driver gathers some board configuration data and
specific environment variables needed for UEFI and Windows IoT. These
are stored in a struct at CFG_FS_DEVICEINFO_ADDR that remains in memory
when UEFI is loaded.
UEFI then uses this data for display initalization among others. It
marks the memory as reserved so it can later be read by Windows.