Add CONFIG_DM_SERIAL to fsimx8mp_defconfig for correct order of TTYs
Without the driver model, the serial driver just iterates through all
serial devices for their number.
The order of the serial devices in the OSM modules get changed in the
device-tree, so activate the driver model to ensure that the order is
correct and that the correct TTY is used in kernel.