Fix saving of fs images on closed boards
When the Board is closed it is unable to save the Bootloader Images
because of two reasons.

The first reason is, that the ELE Firmware is apperently unable to
verify NXP signed Software during runtime. Currently we can only make it
work by skiping the check of the ELE-FW Image, until there is a way to
verify before saving. The ELE is still checked during boot.

The second reason comes from the BOARD-ID. This field is not signed.
During a closed boot, the U-Boot does not accept this unsigned part. We
need to skip this to, since it is not meant to be a signed part of the
Bootloader Image.

Skip the verification of these two Images at saving, so that a closed
Board is able to save new firmware.