Search for TEE while saving independent from CONFIG_IMX_OPTEE
If a user wants to save an OP-TEE enabled N-Boot while using an
U-Boot without CONFIG_IMX_OPTEE enabled it will skip saving the
OP-TEE. This breaks the system during boot because the bl31 build
for jumps into the OP-TEE location where no code will be located.

Change this system to always take OP-TEE into account to not break
on saving.