Add support for variable updategpio when doing updates
During boot, code in common/autoboot.c calls board_check_for_recover() to
check if a regular update or a recover procedure should be started. The
idea is to check if a special recover button is pressed. The user can set
a variable recovergpio with a GPIO number and a state high or low. If
the GPIO has this state, the recover process is triggered.

In the same fashion, the user can now set a variable updategpio. If the
variable is set, then the GPIO must have the given state to trigger the
regular update, otherwise no update takes place. This is meant as an
additional safeguard. An update is only done if the specified button
on this GPIO is also pressed during boot. The existence of the update
script alone is not sufficient then.

If variable updategpio is not set, the behaviour is as before. Which
means it is sufficient that the update script is present to trigger the
update procedure.