Fix NFSv3 mount handling
The mount RPC returns a directory filehandle to be used in the following
NFS_LOOKUP command. This file handle had a fix size in NFSv1/v2 and now
has a variable size including a length entry in NFSv3. This was not
handled correctly. Backport a commit from u-boot-2025.04 that handles
the filehandle depending on the NFS version.