Bring back F&S U-Boot handling in SPL
Before u-boot-2ß24.04, there was a separate image load function for each
boot device (MMC, NAND, SDP, etc.). This was merged to a common function
in include/spl_load.h in u-boot-2024.04.

Bring back the F&S handling of U-Boot images with an optional F&S-Header
prepended. This was lost during the previous merge.