Handle tmu initialization for imx93 and imx8ulp in soc.c for now
For imx93, imx91 and imx8ulp the TMU already is initialized in soc.c!
Double initialization seems to break something in the global data
struct, as disabling the dcache forces a critial reset.
We would either need to move the initialization to fs_board_common as
well or simply mask the call for these architectures.
ATTENTION: Further testing is needed here in any case!!!!