F+S:common:fs_board_common: fix NULL ptr exception
uclass_get_device() can not handle NULL as devp.
It will alwas write the address of the udevice object in devp.
A NULL will cause a NULL-PTR-Exception.

In addition:
uclass_get_device() will only initialize the driver, when it is not
probed and allways returns the udev object of the driver instance.
Multiple calls should have no effects in system stability.