Make F+S comon files compilable outside of U-Boot
In preparation for the fsimage tool for Linux, change some files in
board/F+S/common and include/ to be compilable outside of the U-Boot
build environment.

The files are sill buildable and most changes are added through a
destinction if __UBOOT__ is set or not. This define is set while
beuilding the U-Boot or the SPL.

Some functionality was not ported yet because it needs additional work.
This includes Secure Boot. All places where additional work is needed
where marked with TODOs.

The compilation was tested with the following command on all changed .c
files:

arm-linux-gcc -c <filename> -DUSE_HOSTCC -DCONFIG_FS_CNTR_COMMON
-DCONFIG_MMC -DCONFIG_IMX93 -DCONFIG_FS_BOARD_CFG
-DCONFIG_TARGET_FSIMX93 -DAHAB_BOOT