dts:efusmx93: set broken_cd in fs_sdio_a
There seems to be an issue with card detection (CD) in U-Boot.
This requires further investigation.
The `broken-cd` property is set, which has the same effect as
"non-removable".

As a result, `sdhci_get_cd()` will always return true, which is
acceptable for U-Boot.

It is still possible to switch SD cards on this port.
Reinitialization using `mmc dev <x>` might be required!

This issue affects only efusmx93. OSM93 and PicoCoreMX93 seems to be
fine here.