Mark XHCI in SPL as optional
Please see USB_DWC3 KConfig entry as reference:
The USB_DWC3 driver depends on either USB_XHCI_HCD or USB_GADGET, so
when disabling USB_XHCI_HCD in SPL, the gadget driver can still function
on its own.
This way the phy and driver initialization is handled in our
architechture-specific spl.c like before.
If we ever need host mode in SPL, we need to reenable the XHCI part of
the driver, which will then require the MULTI_DTB solution in SPL.