Fix device-trees for fsimx8mp
ENET:
For imx8mp the eqos driver now aligns more with the Linux driver. Also
move the PHY resets to the PHY node and increase the post-delay for
realtek to 100ms, as the datasheet describes a minimum of 72ms delay
before register calls.

SERIAL:
Correctly set stdout-path and pinctrl for debug UART.

OSM8MP:
Revert some of the UBoot specific changes, as they are now already
handled in imx8mp.dtsi