fsimage.sh: Add full support for i.MX container headers
On F&S boards, all i.MX containers are enhanced by adding F&S image
headers, making them to F&S images. This allows to list the container
contents with the fsimage.sh tool.

The surrounding F&S header uses the extra offset option to skip the
i.MX container header, so the first real image from the F&S image point
of view is the INDEX image. In the past, this extra region could only
be shown as unknown data. Then there was code added to show some
rudimentary info about the container headers in this region. However
this implementation broke the extraction code and also broke the
output format.

Improve the container header code so that this actually works. Now the
i.MX container header in the extra region between the F&S header and
the INDEX is shown like any other image and can be extracted, too.