fsimx93: Have generic container code for NBoot build
Building NBoot on CPUs with i.MX containers depends on two different
sources of information:

1. The Makefile that prepares all image files that are needed in the
   final container.
2. The cfg file for mkimage that holds the list of files actually to be
   included in the container.

In the past, both sources had the list of files. It was essential that
the sequence of files in both lists was exactly the same or the F&S
header (built by the Makefile) and the F&S image (added by the cfg
file via mkimage) would not match. This was prone for errors.

In addition, each container that is part of NBoot needed to be
implemented as own rules in the Makefile, which made the Makefile
quite complex.

This commit now drops this behavior and implements a very clean build
process for containers, only based on the file list in the cfg file.

The new makefile board/F+S/common/fus-common.mk defines generic functions
that can be used for all NBoot variants. Just include this file in a board
specific Makefile and use these functions.

One part consists of cmd entries:

- cmd_fsimg: Call addfsheader for one F&S image
- cmd_copy: Copy a file
- cmd_sed: Call sed on a file
- cmd_cat: Concatenate several files with cat
- cmd_dram_cc: Call compiler for a DRAM-TIMING file
- cmd_dram_elf: Call linker for a DRAM-TIMING file
- cmd_dram_bin: Convert DRAM-TIMING from elf format to binary format
- cmd_cfg: Substitue path variables in a cfg file
- cmd_index: Create an F&S INDEX for several images

The second part is a special template fsc (F&S container). When called
and evaluated somewhere else, this template generates a set of rules
required to build exactly one type of container. For example if NBoot
needs three containers, boot_info.fs, board_info.fs and lpddr4_info.fs,
then you need to evaluate the template three times, once for boot_info,
once for board_info and once for lpddr4_info.

For example when called with argument "board_info", then "board_info" is
used as part of the rules, variables and filenames that are generated by
the template. For example it creates sed code to extract all filenames
from the file "board_info.cfg" and stores them in a variable
"board_info_images". And it creates a rule for target "board_info.fs"
with prerequisites that depend on this variable "board_info_images".
It also creates a rule to build the F&S INDEX and for all calls to
addfsheader.

For this to work, the filenames in the cfg file need to follow a certain
pattern:

  <descr>_<type>.fsi

<type> and <descr> are the values that are passed with options -t and -d
to addfsheader. For example a board configuration may be called

  PCore93-FERT2_BOARD-CFG.fsi

This will create an F&S Header with type BOARD-CFG and description
PCore93-FERT2. The only thing that needs to be defined in the board
specific Makefile is how to build such files from the .dts files by
defining a pattern rule to call the device tree compiler, something
like this:

%_BOARD-CFG.fsi: %.dts
	$(call cmd,dtco)

As a simplification, the paths of the files in a cfg file can be
replaced by path variables:

- $(src) will be replaced by the current directory in the source tree
- $(obj) will be replaced by the current directory in the build directory
- $(nxp) will be replaced by the NXP-Firmware directory

By the way this container concept also works for U-Boot.

The changes of this commit in more detail:

- Add board/F+S/common/fus-common.mk with the above template. The file
  also explains in detail how F&S images and F&S containers look like.
- Add comment to board/F+S/common/fs_cntr_common.c to explain how a
  container based F&S NBoot image looks like.
- Change board/F+S/fsimx93/nboot/Makefile to use the above concept for
  NBoot.
- Replace boot-info.cfg with boot_info.cfg and move to nboot subdir.
- Replace board-info.cfg with board_info.cfg and move to nboot subdir.
- Replace lpddr4_type.cfg with lpddr4_info.cfg and move to nboot subdir.
- Change board/F+S/fsimx93/Makefile to use the above concept for U-Boot.
- In board/F+S/fsimx93, replace bl31-bl32-bl33.cfg and bl31-bl33.cfg
  with uboot.cfg. The TEE line will automatically be deleted if U-Boot
  is configured without opTEE support, so one cfg file is sufficient.
- Drop CONFIG_FUS_BOOTCNTR_CONFIG and CONFIG_FUS_UBOOTCNTR_CONFIG the
  name given in the Makefile is sufficient.
- Change scripts/Makefile.lib so that the device tree compiler also finds
  nboot-info.dtsi if building out of tree.
- Change tools/imx8image.c to print the container image offset as hex number
  preceeded by "0x".
- Make CONFIG_FS_CNTR_COMMON and CONFIG_FS_BOOTROM hidden configs, it does
  not make sense to manually select or deselect them on F&S boards.
- Remove multiple inclusion of board/F+S/common/Kconfig through the board
  specific Kconfigs in fsimx91, fsimx93 and fsimx8ulp.
- Update and fix OCRAM layout in fsimx93.h, add DRAM layout.

Instead of uboot-info.fs, the resulting U-Boot image is called uboot.fs again,
like on i.MX8 boards.

- Drop target prepare_fus from global Makefile, it is no longer needed.
- Drop rules to clean .cntr and .cntr.log, we use .*.tmp files now that are
  automatically cleaned.

Remark:
The modification just changes fsimx93, so fsimx91 and fsimx8ulp will not build
until the changes are also applied to their Makefiles.