Change OCRAM memory layout on fsimx93
After DRAM initialization, function ddr_init() copies the dram timing to
CONFIG_SAVED_DRAM_TIMING_BASE, where it is expected by the ATF. On
regular
NXP boards, this is necessary because the dram timing is part of the SPL
code. However on F&S boards, the dram timing is separate from SPL in a
self-contained F&S image (DRAM-TIMING). Thus it can be loaded directly
to CONFIG_SAVED_DRAM_TIMING_BASE. This saves 16KB of OCRAM memory for
the extra copy and also saves the time for the copy process.

Add option CONFIG_SKIP_DRAM_TIMING_SAVING to skip copying the dram
timing to CONFIG_SAVED_DRAM_TIMING_BASE. This would be an in-place copy
which is a) not necessary and b) would most probably not work anyway,
because data is restructured in function dram_config_save().

Change linker script dram-timings.lds accordingly so that variable
dram_timing is directy at the beginning of the DRAM-TIMING image.
This is how ATF expects it. Also change the loading code in
board/F+S/fsim93/spl.c accordingly, the extra pointer to this
variable is not needed anymore.

Add new 8KB space CFG_FUS_INDEX_ADDR where all INDEX images are loaded
to. Change all mkimage cfg files to accordingly, now they are much more
similar.

Move CFG_SPL_DRAM_FW_ADDR more to the back to close the gap to
CFG_FUS_INDEX_ADDR. This allows SPL to grow larger in the future.