F+S:common:fs_board_common: declare sec_boot environment in late_init
sec_boot should be a generic environment for checking whether the board
is closed. It is set during boot according to the close status flag of
HAB/AHAB.

The environment can be taken into account in the shell, e.g. in boot
scripts, or in the code, e.g. in cmd update.