F+S:common:fus_common: allow cfg files without INDEX entry
The i.MX8ULP BootROM does not support OEM-specific images (e.g., F&S
index files) as the first image within the BOOT container.
If such a data image is placed first, the BootROM enters an error state,
which prevents loading the SPL, UPOWER, and Cortex-M images.

This patch introduces a safer way to handle the optional index entry:

$(1)_index := $(or $(word 1, $(filter %_INDEX.fs, $($(1)_images))),
$(obj)/.$(1)_index.fs)

This line checks whether a matching _INDEX.fs entry exists in the
boot_info.cfg file. If found, the first match is used. If not, a dummy
index is created to ensure the target rule remains valid for GNU Make:

  $($(1)_index): $($(1)_others)
  	$(call cmd,index)

This approach ensures a consistent build process and avoids invalid
container Images that the i.MX8ULP BootROM cannot handle.