configs:fsimx93: add FUS_BOOTCMD
This patch introduces `FUS_BOOTCMD`, an alternative to the existing FuS
environments. Similar to U-Boot’s distro bootcmd, it is based on the
standard boot concept and handles environments for different bootdev,
bootmeth, and bootflow configurations.

A virtual bootdev declaration `bootcmd_fus_legacy` is provided for
compatibility with the old boot standard.

Currently supported devices and features:
- MMC
- USB
- TFTP/NFS
- Secure Boot
- Cortex-M Boot
- A/B Boot

Currently missing:
- NAND/UBIFS