F+S:common:fs_board_common: define generic fs_board_cma_fdt_fixup
fs_board_cma_fdt_fixup can be used in OF_BOARD_SETUP to prepare the cma
size in Device-Tree.

It uses the environment cma_size in U-Boot to determine the size in FDT.

If the environment is not defined and dram size is less then 1GiB, then
the default size in FDT will be used.

Also, if the environment is not defined and DRAM-Size is bigger then
1GiB, then the size will be set to 640MiB.