Fix nest_level in fs_image_common.c
In fs_image_start(), nest_level is initialized to -1. Meaning that the
following call to fs_image_enter() will increase it to 0 to start the
next nesting level. However the call also decreases the remaining size
of the current level, which would reference fsimg_stack[-1]. Fix this
by checking for nest_level >= 0.