VAND3 does not need WLAN MAC from F&S address range
VAND3 uses SDMAC, not SDPAC. On SDMAC, there is a pre-programmed MAC
address and thus the board does not need a MAC address from F&S
address range. Do not call fs_fdt_set_wlan_macaddr() in this case.
Add comment to avoid that it is added back again in the future by
mistake.