F+S:fsimx93: add FUS_FORCE_DEFAULT_BOOTDELAY
If board is closed and Secure-Boot is enabled, then the board will
disable the U-BOOT shell via bootdelay to provide a save Secure-Boot
environment.

The Config FUS_FORCE_DEFAULT_BOOTDELAY can be used to keep the default
bootdelay even if board is closed.

If enabled, bootdelay defined in configs/<board>_defconfig keeps it
value, even during chain of trust. Use this setting only for debug
purpose.