board:F+S:fsimx93: Disable shell access during closed board and fastboot
To prevent access to the SHELL with valid images in combination with
Secure Boot, FASTBOOT rigorously halts the board to continue ensuring a
secure environment.

To enable access to the shell for development or normal bring-ups,
CONFIG_FUS_FORCE_DEFAULT_BOOTDELAY must be set and the corresponding
build must be signed.