usb:gadget:ci_udc: workaround pcore8ulp usb-id
When used with PCoreBBDSI, the ID PIN for USB-OTG does not function
reliably on the PCoreMX8ULP.  Therefore, if the USB-PHY is not
recognised as an INIT_DEVICE in the FDT, it is checked to see if the
dr_mode property has been set to 'peripheral'.

If DR_MODE is defined as a peripheral device, we ignore the status of
the ID pin and assume a device configuration.

With this cange, fastboot commands works reliably.