F+S:fsimx8ulp: add Secure-Boot support
On the i.MX8ULP, it appears that the ELE performs configuration of the
tRDC (setting RWX flags) when data is placed in SSRAM. As a result, the
ELE firmware does not allow overwriting already loaded and validated
image blobs with new valid images. When the ELE revalidates a memory
region already marked in the tRDC, it triggers an internal reset.

This behavior is counterproductive during NBOOT, since the Board-CFG and
DRAM timings must be loaded sequentially depending on the specific board
and manufacturing variations.

To work around this issue, only the signature of the CNTR-HDRs is
validated using the AHAB API. For the i.MX8ULP and newer devices, the
image hashes are verified via software instead.