Add IVT header for SPL before adding NBoot header for imx8m*
This is a fixup for commit 86266a42eb833c21673f4f2966a566789f78fa36.
Because spl/u-boot-spl.bin does not have the IVT header from imx8mimage
included, we add a step to do this in our NBoot Makefile. The
spl/u-boot-spl.cfgout is automatically generated for each build with
scripts from the NXP Makefile. It cannot be handled as a Makefile
target/object.