Force Host Mode for xhci-imx8m
In earlier version of the driver, imx8m_xhci_core_init and
board_usb_init were switched from their calling order. If we get here,
we know that the device should be able to run in Host Mode, so force the
Host Mode as in earlier versions.
The ID pads in hardware can not be used anyway.