Change CFG_FUS_BOARDCFG_ADDR back to CONFIG_FUS_BOARDCFG_ADDR
In the past we changed CONFIG_FUS_BOARDCFG_ADDR to
CFG_FUS_BOARDCFG_ADDR. The Reason is a change to the U-Boot where
Macros from the include files should have the CFG prefix and Macros
from the KCONFIG have the prefix CONFIG. This change caused the
autoconf file, to not have this attribute since it only searches
for CONFIG.

This caused an issue where the fsheader of the BOARD-CFG had its
p32[7] field set to 0 for i.MX8M Boards. This made it impossible
to correctly sign the BOARD-CFG. Running this NBoot as a signed
NBoot broke even open Boards.

Because of this change the MACRO back to the CONFIG prefix.