Fix pointer Arithmetic to copy signed Images
A memcopy command was moved to another function in the past. The
Syntax of the original command used unsigned long for the addresses.
After copying it used struct fs_header_v1_0*. This causes wrong copy
ranges.

Cast the type to uintptr_t to fix this problem and make the necassery
datatype clear.