fsimx8mm/mn/mp: Add extra region for ATF/TEE
In the future, it should be possible to update ATF and TEE separately from
NBoot. This means the current NBoot layout, where ATF and TEE are part of
the FIRMWARE sub-image, is not suited very well anymore.

Move ATF and TEE to a separate ATF region in flash memory. This region is
defined by two new entries atf-start and atf-size in nboot-info.dtsi. The
new property "support-u-atf" indicates the presence of these values.

As the FIRMWARE sub-image now only consists of DRAM-INFO, it can be dropped
completely. From now on, DRAM-INFO, ATF and TEE are directly sub-images of
NBOOT. See layout in fs_image_spl.c.

For quite a while, only one DRAM-FW and one DRAM-TIMING are stored in flash
memory. So even the DRAM-TYPE sub-image is no longer necessary in flash.
This results in the following rather simple flash layout:

    SPL-Region:
   +------------------------------+--- spl-start
   | SPL                          |    spl-size
   +------------------------------+---

    NBOOT-Region
   +------------------------------+--- nboot-start
   | BOARD-CFG (board-id)         |
   +------------------------------+
   | DRAM-FW (type)               |    nboot-size
   +------------------------------+
   | DRAM-TIMING (ram-chip)       |
   +------------------------------+---

    ATF-Region (neu)
   +------------------------------+--- atf-start
   | ATF/U-ATF (arch)             |
   +------------------------------+    atf-size
   | TEE/U-TEE (arch) (optional)  |
   +------------------------------+---

    UBOOT-Region
   +------------------------------+--- uboot-start
   | U-BOOT (arch)                |    uboot-size
   +------------------------------+---

However, the missing group headers that held the total size of the
images in the past make it impossible to still load the images from
flash with the same state machine that is used for SDP-streaming via
USB.

This results in the following changes:

- Change layout description and move from fs_image_common.c to fs_image_spl.c.
- Change nboot Makefiles for fsimx8mm/mn/mp to create the new layout without
  the FIRMWARE sub-image.
- Add support-u-atf, atf-start and atf-size to nboot-info for these
  architectures.
- Enhance fs_image_get_nboot_info_nand/mmc() to also parse these new entries.
- Change fsimage_imx8_save() to handle the new layout in the downloaded
  file as well as in flash memory.
- Change fsimage_imx8_load() to handle the new layout in flash memory.
- Change state machine for SDP streaming to handle the new layout without
  FIRMWARE sub-image.
- Change fs_image_load_system() to load from flash without the state machine.
  Add functions fs_image_load_system_copy() to load a full copy from NBOOT
  and ATF regions and fs_image_load_system_sub() to load and validate one
  sub-image. Remove fs_image_loop() that is no longer needed.